Rename components
- add createRef TODO [ ]
