+++ components/Editor.js
@@ -1,41 +1,55 @@
-// Theirs
 import React from 'react'
-import { Provider } from 'unstated'
-import HTML5Backend from 'react-dnd-html5-backend'
-import { DragDropContext } from 'react-dnd'
+import { Subscribe } from 'unstated'
+import ReadFileDropContainer, { DATA_URL, TEXT } from 'dropperx'
 
-// Ours
 import EditorContainer from '../containers/Editor'
+import Overlay from './Overlay'
 import Coder from './Coder'
-import Toolbar from './Toolbar'
-import { COLORS } from '../lib/constants'
+
+import { DEFAULT_CODE } from '../lib/constants'
+import { isImage } from '../lib/util'
+
+const to = [EditorContainer]
 
 class Editor extends React.Component {
   constructor(props) {
     super(props)
-    this.inject = [new EditorContainer(props)]
+    this.renderPane = this.renderPane.bind(this)
+  }
+
+  renderPane(editor) {
+    return (
+      <ReadFileDropContainer readAs={readAs} onDrop={editor.onDrop}>
+        {({ isOver, canDrop }) => (
+          <Overlay
+            isOver={isOver || canDrop}
+            title={`Drop your file here to import ${isOver ? '✋' : '✊'}`}
+          >
+            <Coder
+              config={editor.state}
+              updateCode={editor.updateCode}
+              onAspectRatioChange={editor.updateAspectRatio}
+              titleBar={editor.state.titleBar}
+              updateTitleBar={editor.updateTitleBar}
+            >
+              {editor.state.code != null ? editor.state.code : DEFAULT_CODE}
+            </Coder>
+          </Overlay>
+        )}
+      </ReadFileDropContainer>
+    )
   }
 
   render() {
-    return (
-      <Provider inject={this.inject}>
-        <div id="editor">
-          <Toolbar />
-          <Coder />
-        </div>
-        <style jsx>
-          {`
-            #editor {
-              background: ${COLORS.BLACK};
-              border: 3px solid ${COLORS.SECONDARY};
-              border-radius: 8px;
-              padding: 16px;
-            }
-          `}
-        </style>
-      </Provider>
-    )
+    return <Subscribe to={to}>{this.renderPane}</Subscribe>
   }
 }
 
-export default DragDropContext(HTML5Backend)(Editor)
+function readAs(file) {
+  if (isImage(file)) {
+    return DATA_URL
+  }
+  return TEXT
+}
+
+export default Editor
